% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePlotter.R
\name{agePlotter}
\alias{agePlotter}
\title{Age plotter}
\usage{
agePlotter(
  y,
  age,
  mod = matrix(rep(1, length(y)), ncol = 1),
  mainText,
  smoothIt = TRUE,
  jitter = TRUE,
  ageLabel = "bottom",
  orderByAge = TRUE,
  ylim = NULL,
  ageBreaks = c(-1, 0, 1, 10, 100),
  ylab = "Adjusted Expression",
  pointColor = 2,
  lineColor = 1,
  alreadyFitted = NULL,
  has_fetal = TRUE,
  ...
)
}
\arguments{
\item{y}{A vector with the Y outcome.}

\item{age}{A vector with the age information.}

\item{mod}{A model matrix.}

\item{mainText}{A character vector with the main title.}

\item{smoothIt}{A logical indicating whether to smooth Y.}

\item{jitter}{A logical indicating whether to add jitter on the X axis.}

\item{ageLabel}{The location of the age legend.}

\item{orderByAge}{A logical indicating whether to sort the observations by
age.}

\item{ylim}{A length two vector with the Y axis limits.}

\item{ageBreaks}{The age cutoffs for the different groups.}

\item{ylab}{The Y axis label.}

\item{pointColor}{An integer indicating the palette color to use for
the points.}

\item{lineColor}{An integer indicating the paletter color to use for the
lines.}

\item{alreadyFitted}{The output of \link[stats:fitted.values]{fitted} on a linear model
if you already calculated it. If so, \code{y} will be ignored.}

\item{has_fetal}{A \code{logical(1)} indicating whether there are any prenatal samples. If so, the first
age window will be shown with the PCW (pre-conception weeks) label and in PCW units.}

\item{...}{Additional parameters to pass to \link[graphics:plot.default]{plot}.}
}
\value{
A nice plot =)
}
\description{
Plots a Y output by age.
}
\details{
\code{pointColor} can be a vector of length equal to \code{age} and have
multiple values in which case \code{lineColor} has to have a length
equal to the number of unique \code{pointColor} values. Specifying
this will draw a line for each unique \code{pointColor}.
}
\examples{


def.par <- par(no.readonly = TRUE) # save default, for resetting...

## Generate some Ys
set.seed(20190827)
y <- as.vector(vapply(0:5, function(x) rnorm(20, mean = x), numeric(20)))

## Generate some ages
age <- as.vector(
    mapply(runif,
        c(-1, 0, 1, 10, 20, 50), c(0, 1, 10, 20, 50, 90),
        n = 20
    )
)

## Make the age plot with a regression line across age vs Y
agePlotter(y, age,
    mainText = "agePlotter example",
    ylab = "Random Ys",
    mod = model.matrix(~age),
    ageBreaks = c(-1, 0, 1, 10, 20, 50, 100)
)

## Example with no prenatal samples
agePlotter(y[age >= 0], age[age >= 0],
    mainText = "agePlotter example",
    ylab = "Random Ys",
    mod = model.matrix(~age[age >= 0]),
    ageBreaks = c(0, 1, 10, 20, 50, 100),
    has_fetal = FALSE
)

## Define line and point colors
p_cols <- rep(rep(c("skyblue3", "dark orange"), each = 10), 6)
l_cols <- c("lightgoldenrod", "light blue")

## Make the plot with the above colors
agePlotter(y, age,
    mainText = "agePlotter example: no prenatal",
    ylab = "Random Ys",
    mod = model.matrix(~age),
    ageBreaks = c(-1, 0, 1, 10, 20, 50, 100),
    pointColor = p_cols,
    lineColor = l_cols
)
legend("bottom", c("DLPFC", "HIPPO"),
    col = l_cols, lwd = 3, bty = "n",
    ncol = 2
)


par(def.par) #- reset to default
}
\author{
Andrew E Jaffe, Leonardo Collado-Torres (examples)
}
