% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corner.R
\name{corner}
\alias{corner}
\title{corner, an improved head()}
\usage{
corner(mat, n = 6)
}
\arguments{
\item{mat}{A DataFrame, data.frame, GRanges, matrix, tibble, or list}

\item{n}{Integer number of rows and columns you want returned}
}
\value{
An n x n object of the original class (or a list of n nxn objects)
}
\description{
Returns the first part of an object similar to head() but subsets
columns as well as rows, and also handles lists of objects.
}
\examples{

m <- matrix(rnorm(1000), ncol = 10)
corner(m)

lis <- list(iris, mtcars, matrix(rnorm(1000), ncol = 10))
corner(lis)

}
\author{
Emily E. Burke
}
