% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getF.R
\name{getF}
\alias{getF}
\title{Given two nested models, calculate the F-statistics}
\usage{
getF(fit, fit0, theData)
}
\arguments{
\item{fit}{An object created with \link[limma:lmFit]{lmFit} using the alternative
model (the larger model).}

\item{fit0}{An object created with \link[limma:lmFit]{lmFit} using the null
model (the smaller model, nested in the larger one).}

\item{theData}{The data used in the \link[limma:lmFit]{lmFit} call.}
}
\value{
A data.frame with the F-statistics (\code{fstat}), the degrees of
freedom for the nallternative model (\code{df1}), the null model
(\code{df0}), and the p-value given the F-distribution (\code{f_pval}).
}
\description{
For some data Y organized in a matrix, calculate a F-statitic per row
comparing two nested models.
}
\details{
This function can also work with outputs from \link[stats:lm]{lm}.
}
\examples{

set.seed(20161005)
## From limma::limFit example page:
sd <- 0.3 * sqrt(4 / rchisq(100, df = 4))
y <- matrix(rnorm(100 * 6, sd = sd), 100, 6)
rownames(y) <- paste('Gene', seq_len(100))
y[1:2, 4:6] <- y[1:2, 4:6] + 4

## Define the alternative and null models
pheno <- data.frame(group = rep(c(0, 1), each = 3), RIN = runif(6) + 8)
mod <- model.matrix(~ pheno$group + pheno$RIN)
mod0 <- model.matrix(~ pheno$RIN)

## Fit the models
library('limma')
fit <- lmFit(y, mod)
fit0 <- lmFit(y, mod0)

## Calculate the F statistics for these nested models
finfo <- getF(fit, fit0, y)
head(finfo)

## You can then use p.adjust() for multiple testing corrections
qvals <- p.adjust(finfo$f_pval, 'fdr')
summary(qvals)

}
\author{
Leonardo Collado-Torres, Andrew E Jaffe
}
