% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOR.R
\name{getOR}
\alias{getOR}
\title{Calculate the odds ratio}
\usage{
getOR(x)
}
\arguments{
\item{x}{A two by two matrix.}
}
\value{
The odds ratio calculated from `x`.
}
\description{
Calculate the odds ratio from a 2 by 2 table.
}
\examples{

getOR(matrix(1:4, ncol = 2))

}
\author{
Andrew E Jaffe
}

