% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPcaVars.R
\name{getPcaVars}
\alias{getPcaVars}
\title{Calculates the percent of variance explained for principal components}
\usage{
getPcaVars(pca, digits = 3)
}
\arguments{
\item{pca}{An object created with \link[stats:prcomp]{prcomp}.}

\item{digits}{The number of significant digits to round to.}
}
\value{
A vector with the percent of variance explained for each of the
principal components in decreasing order.
}
\description{
Given a PCA object created with \link[stats:prcomp]{prcomp}, this function computes
the percent of variance explained by each of the principal components.
}
\examples{

pca <- prcomp(USArrests)
getPcaVars(pca)

}
\author{
Andrew E Jaffe
}
