% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googledrive_csv.R
\name{googledrive_csv}
\alias{googledrive_csv}
\title{Upload a CSV file to googledrive}
\usage{
googledrive_csv(df, sheet_name)
}
\arguments{
\item{df}{A \code{data.frame()} to upload.}

\item{sheet_name}{A \code{character(1)} vector with the name of the Google
Drive spreadsheet that will be updated (or created if it doesn't exist).}
}
\value{
A one row table with the \code{googledrive} information for the file
that was uploaded.
}
\description{
This function uploads a data.frame as a CSV file to Google Drive. If the
name of the spreadsheet exists already, then it gets updated. Otherwise
it uploads it for the first time. Unlike using
\code{drive_upload(overwrite = TRUE)} directly, this function allows you to
keep the sharing and publishing settings of a spreadsheet if it exists
already.
}
\examples{

if(googledrive::drive_has_token()) {
## You'll need to have access to Google Drive through
## googledrive::drive_auth() set up.


## Upload a table to google drive
googledrive_csv(mtcars, paste(Sys.Date(), 'jaffelab::googledrive_csv() example'))

}

}
\author{
Leonardo Collado-Torres
}
