% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucsc_to_granges.R
\name{granges_to_ucsc}
\alias{granges_to_ucsc}
\title{Convert a GRanges object to UCSC coordinates}
\usage{
granges_to_ucsc(gr)
}
\arguments{
\item{gr}{A GRanges object.}
}
\value{
A character vector of UCSC genome browser coordinates.
}
\description{
This function takes a GRanges objects and returns a string of UCSC genome
browser coordinates.
}
\examples{
ucsc_coords <- c('chr1:1000-2000')
granges_to_ucsc(ucsc_to_granges(ucsc_coords))


}
\seealso{
\link{ucsc_to_granges}
}
\author{
Leonardo Collado-Torres
}
