% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaffelab-package.R
\docType{package}
\name{jaffelab-package}
\alias{jaffelab-package}
\title{Commonly used functions by the Jaffe lab}
\description{
The jaffelab contains miscellaneous functions that are commonly used in the
Jaffe lab \url{http://aejaffe.com/}.
}
\author{
Leonardo Collado-Torres <leo.collado@libd.org>
}
\keyword{package}

