% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_functions.R
\name{n_max}
\alias{n_max}
\alias{n_min}
\title{Get more than just the single max or single min value with n_max() and n_min()}
\usage{
n_max(vec, n = 5)

n_min(vec, n = 5)
}
\arguments{
\item{vec}{A vector from which to select the max or min values.}

\item{n}{The number of values to return.}
}
\value{
The n max or min values of your vector.
}
\description{
Returns the n maximum or n minimum values of a vector.
}
\examples{

x <- rnorm(100)
n_max(x)
n_min(x)

}
\author{
Emily E. Burke
}
