% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split0.R
\name{split0}
\alias{split0}
\title{Splits into a list}
\usage{
split0(x)
}
\arguments{
\item{x}{A vector to split.}
}
\value{
A named list according to the unique elements of \code{x} with the
integer indices of those given elements.
}
\description{
Splits a vector (which is casted as a factor) by its elements returning a
named list with the indices for each unique element of the vector.
}
\examples{

split0(letters[1:3])

## With some repeated info
set.seed(20161005)
abc <- sample(letters[1:3], 9, replace = TRUE)
split0(abc)
}
\seealso{
\link[rafalib:splitit]{splitit}
}
\author{
Andrew E Jaffe
}
