% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitit.R
\name{splitit}
\alias{splitit}
\title{Splits into a list}
\usage{
splitit(x)
}
\arguments{
\item{x}{A vector to split.}
}
\value{
A named list according to the unique elements of \code{x} with the
integer indices of those given elements.
}
\description{
Splits a vector by its elements returning a named list with the indices
for each unique element of the vector.
}
\examples{

splitit(letters[1:3])

## With some repeated info
set.seed(20161005)
abc <- sample(letters[1:3], 9, replace = TRUE)
splitit(abc)

}
\author{
Andrew E Jaffe
}
\seealso{
\link{split0}
}

