% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss.R
\name{ss}
\alias{ss}
\title{Wrapper for string split and sapply}
\usage{
ss(x, pattern, slot = 1, ...)
}
\arguments{
\item{x}{Passed to argument \code{x} of \link[base:strsplit]{strsplit}.}

\item{pattern}{Passed to argument \code{split} of \link[base:strsplit]{strsplit}.}

\item{slot}{An integer specifying which element of the resulting list to
return.}

\item{...}{Additional arguments passed to \link[base:strsplit]{strsplit}.}
}
\value{
A vector with the information extracted from \code{x}.
}
\description{
Given a pattern, it splits and returns the slot of interest.
}
\examples{

## Some example info with two variables separated by a _
info <- c('DLPFC_polyA', 'DLPFC_RiboZero')
ss(info, '_', 1)
ss(info, '_', 2)

## Example where we use the ... argument
ss('a.b.c', '.', slot = 1, fixed = TRUE)

}
\author{
Andrew E Jaffe
}
