% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucsc_to_granges.R
\name{ucsc_to_granges}
\alias{ucsc_to_granges}
\title{Convert UCSC coordinates to a GRanges object}
\usage{
ucsc_to_granges(coords)
}
\arguments{
\item{coords}{A character vector of UCSC genome browser coordinates.}
}
\value{
A GRanges object.
}
\description{
This function takes a string of UCSC genome browser coordinates
and returns them as a GRanges object.
}
\examples{
ucsc_coords <- c("chr1:1000-2000")
ucsc_to_granges(ucsc_coords)
}
\seealso{
\link{granges_to_ucsc}
}
\author{
Leonardo Collado-Torres, Andrew E Jaffe
}
