% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePlotter.R
\name{agePlotter}
\alias{agePlotter}
\title{Age plotter}
\usage{
agePlotter(y, age, mod = matrix(rep(1, length(y)), ncol = 1), mainText,
  smoothIt = TRUE, jitter = TRUE, ageLabel = "bottom",
  orderByAge = TRUE, ylim = NULL, ageBreaks = c(-1, 0, 1, 10, 100),
  ylab = "Adjusted Expression", pointColor = 2, lineColor = 1,
  alreadyFitted = NULL, ...)
}
\arguments{
\item{y}{A vector with the Y outcome.}

\item{age}{A vector with the age information.}

\item{mod}{A model matrix.}

\item{mainText}{A character vector with the main title.}

\item{smoothIt}{A logical indicating whether to smooth Y.}

\item{jitter}{A logical indicating whether to add jitter on the X axis.}

\item{ageLabel}{The location of the age legend.}

\item{orderByAge}{A logical indicating whether to sort the observations by
age.}

\item{ylim}{A length two vector with the Y axis limits.}

\item{ageBreaks}{The age cutoffs for the different groups.}

\item{ylab}{The Y axis label.}

\item{pointColor}{An integer indicating the palette color to use for
the points.}

\item{lineColor}{An integer indicating the paletter color to use for the
lines.}

\item{alreadyFitted}{The output of \link[stats:fitted]{fitted} on a linear model
if you already calculated it. If so, \code{y} will be ignored.}

\item{...}{Additional parameters to pass to \link[graphics:plot]{plot}.}
}
\value{
A nice plot =)
}
\description{
Plots a Y output by age
}
\details{
\code{pointColor} can be a vector of length equal to \code{age} and have
multiple values in which case \code{lineColor} has to have a length
equal to the number of unique \code{pointColor} values. Specifying
this will draw a line for each unique \code{pointColor}.
}
\examples{

}
\author{
Andrew E Jaffe
}
