% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaningY.R
\name{cleaningY}
\alias{cleaningY}
\title{Regress out SVs or PCs}
\usage{
cleaningY(y, mod, P = ncol(mod))
}
\arguments{
\item{y}{The outcome}

\item{mod}{A model matrix}

\item{P}{The number of SVs or PCs to protect based on the column order.
For example, \code{P=2} would keep the intercept term and a case vs diagnosis
term in a model that is ~ Dx + more covariates.}
}
\value{
An object of the same type as `y` with the SVs/PCs regressed out.
}
\description{
Regress out surrogate variables or principal components
}
\examples{


}
\author{
Andrew E Jaffe
}
