% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge.pvaue.R
\name{edge.pvalue}
\alias{edge.pvalue}
\title{Function to compute empirical p-values}
\usage{
edge.pvalue(stat, stat0, pool = TRUE)
}
\arguments{
\item{stat}{vector of observed statistics}

\item{stat0}{matrix of null statistics, where rows are features and columns are iterations}

\item{pool}{features are interchangeable, e.g. unlist all null statistics}

\item{...}{Additional arguments passed to \link[base]{strsplit}.}
}
\value{
A vector of empirical p-values.
}
\description{
Given a pattern, it splits and returns the slot of interest.
}
\examples{


}
\author{
Jeffrey T Leek
}
