% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getR2.R
\name{getR2}
\alias{getR2}
\title{Calculate R squared from a matrix}
\usage{
getR2(p, mod, mod0 = NULL)
}
\arguments{
\item{p}{A matrix that will be passed to \link[limma:lmFit]{lmFit} \code{object}.}

\item{mod}{A model matrix for the alternative model (the larger one).}

\item{mod0}{A model matrix for the null model (the smaller one). If \code{NULL}
then \code{p} will be used to calculate the residual sum of squares of the
null model.}
}
\value{
A data.frame with the R squared and the adjusted R squared.
}
\description{
Using limma, calculate the R squared from a matrix using nested models
}
\examples{

}
\author{
Andrew E Jaffe
}
