% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getT.R
\name{getT}
\alias{getT}
\title{Get t-statistics from limma objects}
\usage{
getT(lmFitObj, coiIndex = 2L)
}
\arguments{
\item{lmFitObj}{A \link[limma]{lmFit} object.}

\item{coiIndex}{An integer specifying the coefficent of interest in the model
for which to get the t-statistics for.}
}
\value{
A two column matrix with the t-statistics and the p-values of the
coefficient of interest.
}
\description{
Get the t-statistics from a \link[limma]{lmFit} object.
}
\examples{

## From lmFit() example page
sd <- 0.3*sqrt(4/rchisq(100,df=4))
y <- matrix(rnorm(100*6,sd=sd),100,6)
rownames(y) <- paste("Gene",1:100)
y[1:2,4:6] <- y[1:2,4:6] + 2
design <- cbind(Grp1=1,Grp2vs1=c(0,0,0,1,1,1))

# Ordinary fit
library('limma')
fit <- lmFit(y,design)
## Get Ts
getT(fit)

}
\author{
Andrew E Jaffe
}
