% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaffelab-package.R
\docType{package}
\name{jaffelab-package}
\alias{jaffelab}
\alias{jaffelab-package}
\title{jaffelab: Commonly used functions by the Jaffe lab}
\description{
The jaffelab contains miscellaneous functions that are commonly used in the Jaffe lab http://aejaffe.com/.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LieberInstitute/jaffelab}
  \item Report bugs at \url{https://github.com/LieberInstitute/jaffelab/issues}
}

}
\author{
\strong{Maintainer}: Leonardo Collado-Torres \email{lcolladotor@gmail.com} (\href{https://orcid.org/0000-0003-2140-308X}{ORCID})

Authors:
\itemize{
  \item Andrew E. Jaffe \email{andrew.jaffe@libd.org} (\href{https://orcid.org/0000-0001-6886-1454}{ORCID})
  \item Emily E. Burke \email{emily.burke@libd.org}
}

}
\keyword{internal}
